wip(client): Page分解2
添加到了Manifest
