+++ Surviver_client/app/src/main/java/com/example/myapplicationtest1/page/LoginInputPage.java
@@ -1,0 +1,13 @@
+package com.example.myapplicationtest1.page;
+
+import android.os.Bundle;
+
+import com.example.myapplicationtest1.R;
+
+public class LoginInputPage extends Page {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.login_input);
+    }
+}
